Move and rename the main Captive Web View error class, and replace use of URLSchemeTaskError with the main error class, and move some other code into a separate file and extension
